requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.3
django-extensions==3.1.5
django-import-export==2.7.1
django-silk==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==9.8.3
gprof2dot==2021.2.21
idna==3.3
ipython==7.29.0
jedi==0.18.1
Jinja2==3.0.3
kombu==5.2.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
